Remove unnecessary <form action> toggling for delete buttons
Now, the <form method> is automatically set to POST once, when
the script is activated, and stays that way from then on. The
original mode (for static pages) is set to GET to make use of the
'graceful fallback' delete confirm functionality in case JavaScript
is disabled or scripts fail to load for some other reason.