Update navbar to the new design
The navbar is updated with the new design and main logo. The main sidebar menu is also included, with non-functional links.Also included is reboot.css file, which is the CSS reset shipped with Bootstrap. Font is updated to Nato Sans. New sections are added to style.css and script.js files. Fixes #45.