Display articles on the search page
This is a temporary measure; it displays articles on the search
page but does not gel with the pagination. More work needs to be
done on the questions end, since a lot of the views assume that
the displayed content type is always questions, wherease in reality
that is no longer the case.